Eliminated compilation warning.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@18674 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/renderer/common/vbufmgr.cpp
@@ -82,7 +82,7 @@ csVertexBufferManager::~csVertexBufferManager()
 
   /// go through list of registered client and tell 'em we are no longer
   /// available
-  for (int i=0; i < vClients.Length (); i++)
+  for (size_t i=0; i < vClients.Length (); i++)
     vClients[i]->ManagerClosing ();
     
   SCF_DESTRUCT_IBASE();
